requirements.txt
torch
transformers
torch
ultralytics 
pillow  
tensorflow
sahi
python-multipart
fastapi
uvicorn
gunicorn